redid movement, added character model
added some free models and animations to the project. also added a background to the encyclopedia.

i redid the movement for the 47832th time because i wanted to use the new character controller and input system. followed a tutorial by iheartgamedev on youtube.

camera is now extra janky because it doesn't follow the player, and jump currently not implemented. i haven't deleted the old character oblong yet.

still, yay progress!